make TestMap a subclass of Map
